cleanup data query classes
